refactor: Updated the import path to reflect changes to the module its depending on.
